refactor: change the name of components and imports
