+++ src/components/Cards/CardUser.tsx
@@ -1,6 +1,6 @@
 import { Client, Technician } from "@/types";
 import { FiEdit3 } from "react-icons/fi";
-import DetailModal from "../detailModal/DetailModal";
+import DetailModal from "../detailModal/userModal";
 
 type CardProps = {
   item: Technician | Client;
@@ -8,7 +8,6 @@ type CardProps = {
 };
 
 const Card = ({ item, classname }: CardProps) => {
-
   return (
     <div
       className={`flex justify-between h-16 text-white rounded-2xl border-2 items-center w-full border-[#e9ecef7b] md:py-5 py-3 md:px-7 px-4 bg-secondaryColor ${classname}`}
@@ -21,7 +20,10 @@ const Card = ({ item, classname }: CardProps) => {
         <p className="text-sm font-medium md:w-1/3 w-1/2">{item?.phone}</p>
       </div>
       <div className="flex gap-2 h-7 items-center">
-        <DetailModal icon={<FiEdit3 className="w-4 h-4" />} user={item as Client} />
+        <DetailModal
+          icon={<FiEdit3 className="w-4 h-4" />}
+          user={item as Client}
+        />
       </div>
     </div>
   );
